Finished setting up HC05 as Master and as Slave and then making them communicate with each other.  Also set up buzzing on the master side when slave detects human presence.  Need to work on the PIR sensor's accuracy and precision. It is not working consistently. Also need to add debouncing, timers and other suppliments.
Finished setting up HC05 as Master and as Slave and then making them communicate with each other.

Also set up buzzing on the master side when slave detects human presence.

Need to work on the PIR sensor's accuracy and precision. It is not working consistently.
Also need to add debouncing, timers and other suppliments.